# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API Documentation",
        "description": "Swagger OpenApi implementation for EduFlow",
        "contact": {
            "email": "support@eduflow.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Liste de tous les cours disponibles",
                "operationId": "02eeda67042dba2d22a517f26fb8eaf1",
                "responses": {
                    "200": {
                        "description": "Liste des cours récupérée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cours"
                ],
                "summary": "Créer un nouveau cours (Enseignant uniquement)",
                "operationId": "a4d6b7e7be59571034428ffbb86aaf0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel Avancé"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Maîtrisez le Repository Pattern"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 49.99
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cours créé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/courses/{id}": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Détails d'un cours spécifique",
                "operationId": "854f6bd61994988360cf069d565b60ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du cours"
                    },
                    "404": {
                        "description": "Cours non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{id}": {
            "put": {
                "tags": [
                    "Cours"
                ],
                "summary": "Modifier un cours existant",
                "operationId": "4d00e54a4b57b3259c40894ad2d0c76b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cours mis à jour"
                    },
                    "403": {
                        "description": "Vous n'êtes pas l'auteur de ce cours"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cours"
                ],
                "summary": "Supprimer un cours",
                "operationId": "3a2888333ed2868b6102094f5348520e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cours supprimé"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recommendations": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Obtenir des recommandations basées sur les intérêts",
                "operationId": "20f885bed41b24184477162e57df2d4c",
                "responses": {
                    "200": {
                        "description": "Liste des recommandations"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{id}/favorite": {
            "post": {
                "tags": [
                    "Cours"
                ],
                "summary": "Ajouter ou retirer un cours des favoris",
                "operationId": "1a7db1cb0d99d367edca136d4fce6750",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut du favori mis à jour"
                    },
                    "404": {
                        "description": "Cours introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Liste de mes cours favoris",
                "operationId": "e9cab0da1126d4b54fe25df1a7eb8071",
                "responses": {
                    "200": {
                        "description": "Liste des favoris récupérée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}